+++ packages/Zooniverse-React-Components/test/user-navigation.test.js
@@ -4,30 +4,18 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import ZooHeader from '../src/components/layout/zoo-header';
+import UserNavigation from '../src/components/layout/user-navigation';
 
-const mainHeaderNavList = [
-  <a href="#">a link</a>
-];
-
-describe('<ZooHeader />', function() {
+describe('<UserNavigation />', function() {
   let wrapper;
   before(function() {
-    wrapper = shallow(<ZooHeader mainHeaderNavList={mainHeaderNavList} />);
+    wrapper = shallow(<UserNavigation />);
   });
 
   it('renders without crashing', function() {});
 
   it('renders grommet components', function() {
-    expect(wrapper.find('Header')).to.have.lengthOf(1);
+    expect(wrapper.find('Anchor')).to.have.lengthOf(2);
     expect(wrapper.find('Menu')).to.have.lengthOf(1);
-  });
-
-  it('renders <ZooniverseLogo />', function() {
-    expect(wrapper.find('ZooniverseLogo')).to.have.lengthOf(1);
-  });
-
-  it('renders nav items', function() {
-    expect(wrapper.find('a')).to.have.lengthOf(1);
   });
 });